Update data_update.py for removing whitespace in credit ocr result (#441)
* Update data_update.py for removing whitespace in credit ocr result

sometimes ocr provides a str like '1 2261000' and conversion to int int('1 2261000') fails. we do a post process here, removing the whitespace and allow the conversion to be successful

* Update data_update.py for removing whitespace in credit ocr result 

sometimes ocr provides a str like '1 2261000' and conversion to int int('1 2261000') fails. we do a post process here, removing the whitespace and allow the conversion to be successful
edited to align with standards of this repo